RecipeBackend/food_recommendation_backend/requirements.txt
annotated-types==0.7.0
anyio==4.8.0
asgiref==3.8.1
cachetools==5.5.2
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
cryptography==44.0.0
Django==5.1.2
django-allauth==65.3.0
django-cors-headers==4.6.0
django-environ==0.11.2
django-filter==25.1
django-nested-admin==4.1.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
exponent_server_sdk==2.1.0
filetype==1.2.0
google-ai-generativelanguage==0.6.16
google-api-core==2.24.1
google-auth==2.38.0
googleapis-common-protos==1.69.1
greenlet==3.1.1
grpcio==1.70.0
grpcio-status==1.70.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.20
langchain-core==0.3.43
langchain-google-genai==2.0.11
langchain-text-splitters==0.3.6
langgraph==0.3.5
langgraph-checkpoint==2.0.18
langgraph-prebuilt==0.1.2
langgraph-sdk==0.1.55
langsmith==0.3.13
msgpack==1.1.0
orjson==3.10.15
packaging==24.2
pillow==11.0.0
proto-plus==1.26.0
protobuf==5.29.3
psycopg==3.2.4
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
PyJWT==2.10.1
python-monkey-business==1.1.0
PyYAML==6.0.2
razorpay==1.4.2
requests==2.32.3
requests-toolbelt==1.0.0
rsa==4.9
setuptools==78.0.2
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.38
sqlparse==0.5.1
tenacity==9.0.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
whitenoise==6.9.0
zstandard==0.23.0
